Move price displayment methods to order detail presenter
