+++ app/presenters/order_detail_presenter.rb
@@ -3,9 +3,10 @@
 class OrderDetailPresenter < SimpleDelegator
 
   include ActionView::Helpers::NumberHelper
+  include ActionView::Helpers::TagHelper
+
   include DateHelper
   include Rails.application.routes.url_helpers
-  include PriceDisplayment
 
   delegate :admin_editable?, to: :reservation, prefix: true
   delegate :template_result, to: :stored_files, prefix: true
@@ -53,12 +54,91 @@ class OrderDetailPresenter < SimpleDelegator
     order_status&.name
   end
 
+  def display_cost
+    number_to_currency(actual_cost) || number_to_currency(estimated_cost) || empty_display
+  end
+
+  def display_subsidy
+    return unless has_subsidies?
+    number_to_currency(actual_subsidy) || number_to_currency(estimated_subsidy) || empty_display
+  end
+
+  def actual_or_estimated_total
+    actual_total || estimated_total
+  end
+
+  def display_total
+    number_to_currency(actual_or_estimated_total) || empty_display
+  end
+
+  def wrapped_cost
+    content_tag :span, display_cost, class: display_cost_class
+  end
+
+  def wrapped_subsidy
+    content_tag :span, display_subsidy, class: display_cost_class
+  end
+
+  def wrapped_total
+    content_tag :span, display_total, class: display_cost_class
+  end
+
+  def display_cost_class
+    if actual_cost
+      "actual_cost"
+    elsif estimated_cost
+      "estimated_cost"
+    else
+      "unassigned_cost"
+    end
+  end
+
+  def actual_cost?
+    actual_cost.present?
+  end
+
+  def wrapped_quantity
+    build_quantity_presenter.html
+  end
+
+  def csv_quantity
+    build_quantity_presenter.csv
+  end
+
   private
 
   # Is a fulfilled order detail nearing the end of the 90 day reconcile period?
   # Returns true if it is 60+ days fulfilled, false otherwise
   def reconcile_warning?
     !reconciled? && fulfilled_at.present? && fulfilled_at < 60.days.ago
+  end
+
+  def build_quantity_presenter
+    quantity_to_display = if product.try(:daily_booking?)
+                            daily_booking_quantity
+                          elsif time_data.try(:actual_duration_mins) && time_data.actual_duration_mins.to_i > 0
+                            time_data.actual_duration_mins
+                          elsif time_data.try(:duration_mins)
+                            time_data.duration_mins
+                          else
+                            quantity
+                          end
+
+    QuantityPresenter.new(product, quantity_to_display)
+  end
+
+  def daily_booking_quantity
+    if time_data.try(:actual_duration_days) && time_data.actual_duration_days.to_i > 0
+      time_data.actual_duration_days
+    elsif time_data.try(:duration_days)
+      time_data.duration_days
+    else
+      quantity
+    end
+  end
+
+  def empty_display
+    "Unassigned"
   end
 
 end
