+++ app/views/order_management/order_details/_costs.html.haml
@@ -1,4 +1,3 @@
-- f.object.send(:extend, PriceDisplayment)
 - price_edit_disabled = actual_cost_edit_disabled? || cannot?(:adjust_price, f.object) || f.object.product.skip_order_review?
 - if f.object.actual_cost?
   .cost-table
